chore: move some assets to static
