--- src/assets/video/meshery/meshmap/touch
@@ -1,1 +1,0 @@
-
