+++ src/sections/Company/Brand/Brand-components/layer5.js
@@ -6,10 +6,9 @@ import { Row, Col } from "../../../../reusecore/Layout";
 import Button from "../../../../reusecore/Button";
 import ColorBox from "../../../../components/ColorBox";
 import { FiDownloadCloud } from "@react-icons/all-files/fi/FiDownloadCloud";
-import Layer5WhiteBg from "../../../../assets/images/layer5/layer5-only/png/layer5-white-bg.webp";
+import Layer5WhiteBg from "../../../../../static/brand/layer5/layer5-only/png/layer5-white-bg.webp";
 import Bookmarks from "../../../../assets/images/layer5/bookmarks.webp";
-import BookmarksPDF from "../../../../assets/brand/bookmarks.pdf";
-import Layer5Icon from "../../../../assets/images/layer5/5 icon/svg/light/5-light-no-trim.svg";
+import Layer5Icon from "../../../../../static/brand/layer5/5 icon/svg/light/5-light-no-trim.svg";
 
 const Layer5Wrapper = styled.section`
     @media (max-width:575px){
@@ -123,7 +122,7 @@ const Layer5Brand = () => {
               Media available for print.
             </p>
             <Row className="bookmarks">
-              <Link to={BookmarksPDF}>
+              <Link to="/brand/bookmarks.pdf">
                 <img className="bookmarks" src={Bookmarks} alt="Layer5 and Meshery Bookmarks" />
               </Link>
             </Row>
